create a new folder for pyhon scripts
